refactor: 💡 dirs on markdown
