+++ src/app/(contents)/contents/layout.tsx
@@ -1,6 +1,6 @@
 import { Header } from "@/components/nav/header";
 import { PAGE_NAME } from "@/constants";
-import { RootTab } from "@/features/contents/components/root-tab";
+import { RootTab } from "@/features/markdown/components/root-tab";
 import type { Metadata } from "next";
 import type { ReactNode } from "react";
